tabungan(feat): repair Component, Page and Props with typescript any from mastah iyan
