+++ next.config.mjs
@@ -1,17 +1,4 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {};
 
-// module.exports = {
-//   basePath: './src',
-// }
-
 export default nextConfig;
-
-// @ts-check
-// /**
-//  * @type {import('next').NextConfig}
-//  **/
-// const nextConfig = {
-//     /* config options here */
-//   }
-//   module.exports = nextConfig
