+++ src/app/dashboard/tabungan/ModalTopUp.tsx
@@ -1,6 +1,4 @@
 "use client";
-
-import { useState, useEffect } from "react";
 import { MdClose } from "react-icons/md";
 
 type KeySantri = {
@@ -15,59 +13,11 @@ type KeySantri = {
   no_tabungan: string;
 };
 
-export default function AddSaldo(props: { id: string }) {
-  // if i add AddSaldo({tabItem: KeySantri}) in props, tabItem is not defined
-  // AddSaldo(id: any) cant used
-  // {id} error but props can readed and used
-  const [openNewTab, setOpenNT] = useState<boolean>(false);
-  const [dataForTopup, setDataForTopup] = useState<KeySantri>();
-  
-  async function getTabSantri() {
-    const res = await fetch(`http://localhost:5000/santri/${props.id}`, { cache: 'no-store'});
-    console.log("123 res", res)
-    return res.json();
-    // to get value from localhost:5000 you must run the json-server on terminal with
-    // npm run server
-  }
-
-  const showmodal = async () => {
-    setOpenNT(!openNewTab)
-    // if (openNewTab === false) {
-    //   const newData = await getTabSantri();
-    //   setDataForTopup(newData);
-    //   setOpenNT(true);
-    //   console.log("1a", newData);
-    //   console.log("1b", dataForTopup);
-    // } else {
-    //   setOpenNT(false);
-    //   setDataForTopup(undefined);
-    //   console.log("2", dataForTopup);
-    // }
-  };
-
-  useEffect(async () => {
-    if (openNewTab) {
-      const newData = await getTabSantri();
-      setDataForTopup(newData);
-    } else {
-      setDataForTopup(undefined);
-    }
-  }, [openNewTab]);
-
+export default function AddSaldo({item, onClose} :any) {
+  console.log(item)
   return (
     <section>
-      {/* <!-- Modal toggle --> */}
-      <button
-        data-modal-target="default-modal"
-        data-modal-toggle="default-modal"
-        className="bg-blue-600 hover:bg-blue-700  text-white focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm px-5 py-2.5 me-2 mb-2"
-        type="button"
-        onClick={showmodal}
-      >
-        Top Up 💰
-      </button>
-      {/* <!-- Main modal --> */}
-      <div className={openNewTab === false ? "hidden" : "block"}>
+      <div>
         <div
           id="default-modal"
           aria-hidden="true"
@@ -79,20 +29,20 @@ export default function AddSaldo(props: { id: string }) {
               {/* <!-- Modal header --> */}
                 <div className="flex items-center justify-between p-4 md:p-5 border-b rounded-t">
                   <h3 className="text-xl font-semibold text-gray-900">
-                    Top Up Tabungan of {dataForTopup?.nama}
+                    Top Up Tabungan of {item?.nama}
                   </h3>
                   <button
                     type="submit"
                     className="text-gray-700 bg-transparent hover:bg-gray-200 rounded-lg text-sm w-8 h-8 ms-auto inline-flex justify-center items-center"
                     data-modal-hide="default-modal"
-                    onClick={showmodal}
+                    onClick={onClose}
                   >
                     <MdClose />
                     <span className="sr-only">Close modal</span>
                   </button>
                 </div>
                 {/* <!-- Modal body --> */}
-                {!!dataForTopup ? (
+                {!!item ? (
                   <form>
                     <div className="max-w-sm mx-auto py-5">
                       <div className="mb-3">
@@ -102,7 +52,7 @@ export default function AddSaldo(props: { id: string }) {
                         <input
                           type="text"
                           id="username"
-                          defaultValue={dataForTopup?.nama}
+                          defaultValue={item?.nama}
                           className="shadow-sm bg-gray-50 border border-blue-400 text-gray-900 text-base rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2"
                           placeholder="nama"
                           required
@@ -115,7 +65,7 @@ export default function AddSaldo(props: { id: string }) {
                         <input
                           type="text"
                           id="no_tabungan"
-                          defaultValue={dataForTopup?.no_tabungan}
+                          defaultValue={item?.no_tabungan}
                           className="shadow-sm bg-gray-50 border border-blue-400 text-gray-900 text-base rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2"
                           placeholder="nim"
                           required
@@ -128,7 +78,7 @@ export default function AddSaldo(props: { id: string }) {
                         <input
                           type="number"
                           id="nim"
-                          defaultValue={dataForTopup?.nim}
+                          defaultValue={item?.nim}
                           className="shadow-sm bg-gray-50 border border-blue-400 text-gray-900 text-base rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2"
                           placeholder="nim"
                           required
@@ -157,7 +107,7 @@ export default function AddSaldo(props: { id: string }) {
                     </div>
                   </form>
                 ) : (
-                  <div className="text-2xl text-red-600 text-bold py-5 px-4 text-center">error dataForTopup undefined</div>
+                  <div className="text-2xl text-red-600 text-bold py-5 px-4 text-center">error item undefined</div>
                 )}
             </div>
           </div>
