+++ src/types/santri.ts
@@ -1,0 +1,12 @@
+
+export type KeySantri = {
+  nama: string;
+  nim: number;
+  orangtua: string;
+  id_kelas: number;
+  alamat: string;
+  id_card: number;
+  saldo: number;
+  id: string;
+  no_tabungan: string;
+};
