+++ tsconfig.json
@@ -18,7 +18,7 @@
       }
     ],
     "paths": {
-      "./*": ["./src/*"]
+      "@/*": ["./src/*"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
